同步性更新
0. 合并PWM库
1. 更新部分Demo 完善部分逻辑

+++ STM32/Src/App/ArmControl/ArmControl.h
@@ -47,7 +47,7 @@ ArmControlResult_t SetOpenLoopClawPosition(uint16_t RotationAngle, uint16_t Axia
  * @brief: Get claw position in **Polar coordinates** in Open loop control system.
  * @param: Pointer of parameters.
  */
-void GetOpenLoopClawPosition(float *RotationAngle, float *AxialLength, float *Z_AxisHeight);
+void GetOpenLoopClawPosition(uint16_t *RotationAngle, uint16_t *AxialLength, uint16_t *Z_AxisHeight);
 
 typedef enum
 {
