+++ STM32/Src/App/Examples/PlaceApple.c
@@ -20,7 +20,7 @@
  */
 void PlaceApple(void)
 {
-	float rotationAngle, axialLength, zAxisHeight;
+	uint16_t rotationAngle, axialLength, zAxisHeight;
 	GetOpenLoopClawPosition(&rotationAngle, &axialLength, &zAxisHeight);
 
 	if(zAxisHeight < ApproachHeight)
