+++ STM32/Src/App/Examples/WashApple.c
@@ -15,7 +15,7 @@
  */
 void WashApple(void)
 {
-	float rotationAngle, axialLength, zAxisHeight;
+	uint16_t rotationAngle, axialLength, zAxisHeight;
 	GetOpenLoopClawPosition(&rotationAngle, &axialLength, &zAxisHeight);
 
 	if(zAxisHeight < WashAppleHeight)
